requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
idna==2.10
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
Serializer==0.2.1
simplejson==3.17.2
sqlparse==0.4.1
urllib3==1.26.3
